stash fixture changes
