+++ optimism/client_blocks.go
@@ -13,9 +13,47 @@ import (
 	"github.com/ethereum-optimism/optimism/l2geth/core/types"
 )
 
+// toBlockNumArg returns a jsonrpc string identifier for a block.
+// If the provided integer is nil, the latest block is returned.
+func toBlockNumArg(number *big.Int) string {
+	if number == nil {
+		return "latest"
+	}
+	pending := big.NewInt(-1)
+	if number.Cmp(pending) == 0 {
+		return "pending"
+	}
+	return hexutil.EncodeBig(number)
+}
+
+// blockByNumber retrieves a block by a given number from the internal client.
+// If index is nil, the latest block is retrieved.
+func (ec *Client) blockByNumber(
+	ctx context.Context,
+	index *int64,
+	showTxDetails bool,
+) (map[string]interface{}, error) {
+	var blockIndex string
+	if index == nil {
+		blockIndex = toBlockNumArg(nil)
+	} else {
+		blockIndex = toBlockNumArg(big.NewInt(*index))
+	}
+
+	r := make(map[string]interface{})
+	err := ec.c.CallContext(ctx, &r, "eth_getBlockByNumber", blockIndex, showTxDetails)
+	if err == nil {
+		if r == nil {
+			return nil, ethereum.NotFound
+		}
+	}
+
+	return r, err
+}
+
 // Block returns a populated block at the *RosettaTypes.PartialBlockIdentifier.
 // If neither the hash or index is populated in the *RosettaTypes.PartialBlockIdentifier,
-// the current block is returned.
+// the current (aka latest) block is returned.
 func (ec *Client) Block(
 	ctx context.Context,
 	blockIdentifier *RosettaTypes.PartialBlockIdentifier,
