+++ optimism/testdata/goerli_bedrock_block_5003318.json
@@ -18,8 +18,47 @@
   "timestamp": "0x63dd1ad0",
   "totalDifficulty": "0x0",
   "transactions": [
-    "0x035437471437d2e61be662be806ea7a3603e37230e13f1c04e36e8ca891e9611",
-    "0x6103c9a945fabd69b2cfe25cd0f5c9ebe73b7f68f4fed2c68b2cfdd8429a6a88"
+    {
+      "blockHash": "0x4503cbd671b3ca292e9f54998b2d566b705a32a178fc467f311c79b43e8e1774",
+      "blockNumber": "0x4c5836",
+      "from": "0xdeaddeaddeaddeaddeaddeaddeaddeaddead0001",
+      "gas": "0x8f0d180",
+      "gasPrice": null,
+      "hash": "0x035437471437d2e61be662be806ea7a3603e37230e13f1c04e36e8ca891e9611",
+      "input": "0x015d8eb900000000000000000000000000000000000000000000000000000000008097790000000000000000000000000000000000000000000000000000000063dd1a98000000000000000000000000000000000000000000000000000000000004ee2f1ed96835176d084c845bd2c09456d60401d74861b690bdabac97f6724f4b4bdf00000000000000000000000000000000000000000000000000000000000000020000000000000000000000007431310e026b69bfc676c0013e12a1a11411eec9000000000000000000000000000000000000000000000000000000000000083400000000000000000000000000000000000000000000000000000000000f4240",
+      "isSystemTx": true,
+      "mint": "0x0",
+      "nonce": "0x0",
+      "r": null,
+      "s": null,
+      "sourceHash": "0xe498acd8ac4c577ba87349e5f649034404485515ba7f2fa3b8dfda726dd62c16",
+      "to": "0x4200000000000000000000000000000000000015",
+      "transactionIndex": "0x0",
+      "type": "0x7e",
+      "v": null,
+      "value": "0x0"
+    },
+    {
+      "accessList": [],
+      "blockHash": "0x4503cbd671b3ca292e9f54998b2d566b705a32a178fc467f311c79b43e8e1774",
+      "blockNumber": "0x4c5836",
+      "chainId": "0x1a4",
+      "from": "0xe261e28d9fccd3742629fef031e63327585b40f0",
+      "gas": "0x5b8d80",
+      "gasPrice": "0xb2d05e61",
+      "hash": "0x6103c9a945fabd69b2cfe25cd0f5c9ebe73b7f68f4fed2c68b2cfdd8429a6a88",
+      "input": "0xb1dc65a40001b9ada1cc34d3d18c4f9705f77b5036df2e9041c9b16c1e511d3dff17ab81000000000000000000000000000000000000000000000000000000000029a1030dabb9edf2d1abbfd18a5f5b5dd8f6fe9e3cac59160d012ad1ad2c312acb741700000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000000000062000000000000000000000000000000000000000000000000000000000000006a00101000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000052000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000120000000000000000000000000000000000000000000000000000000000000016000000000000000000000000000000000000000000000000000000000000001a000000000000000000000000000000000000000000000000000000000000001e0000000000000000000000000000000000000000000000000000000000000026000000000000000000000000000000000000000000000000000000000000004c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004e0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000e6e0000000000000000000000000000000000000000000000000000000000000001000000000000000000000000dc2cc710e42857672e7907cf474a69b63b93089f00000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000f43fc2c04ee00000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000dc2cc710e42857672e7907cf474a69b63b93089f000000000000000000000000000000000000000000000000000000000000a869000000000000000000000000000000000000000000000000000000073890e54a0000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000a8690000000000000000000000000000000000000000000000000000000000000e6e0000000000000000000000000000000000000000000000000000000000000064000000000000000000000000620a71123c7090c9e66daea5235872b250f3c2610000000000000000000000000000000000000000000000000000000000000e6e0000000000000000000000000000000000000000000000000000000000030d400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000e60cead5fcd752b6694f90a16af7a46e5b6df817000000000000000000000000000000000000000000000000000000000000018000000000000000000000000000000000000000000000000000000000000001c00000000000000000000000000b9d5d9136855f6fec3c0993fee6e9ce8a2978466f4086b5b683bcade1f3f98b34e8b3d95c103324a5cd2762619d86b689222c9000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000001cdc000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000397fce5bbf0a173bc6489b5d80dc6f25abe0583d301b3d49af106fb8b7c8b1cb45e601f60295cc3d24785ee8a9677d4a6adeba6bbc452ba9f23e5f06d85bdb936d65c22dc9ce5b03961de239adc29a90d2a8d91a92378bcdd9c684b317c9d78f100000000000000000000000000000000000000000000000000000000000000034474147f9f712b068497b1895ae6b62969d7d29b39a898708028aa6494c8a8eb1df987ff11c5069eed5ca8db0fa5b4ccce7bf7cd2be9a68988194366745e6cd337b633327318d96df91aa35e7330413955593139f40f4c630a179bdf8f03bc94",
+      "maxFeePerGas": "0xb2d05e7e",
+      "maxPriorityFeePerGas": "0xb2d05e30",
+      "nonce": "0x1fc9",
+      "r": "0xf251114b84dbab64fb9629e2298252b09077080169dd3970b7ac06bd73be5a73",
+      "s": "0x7a7481ae00acfa76673656a93b538d1be1bc7b4ecd16204e8e80cc369b2fd63c",
+      "to": "0x794c23bb0a718f4a79ee96531d40c54a67f7f037",
+      "transactionIndex": "0x1",
+      "type": "0x2",
+      "v": "0x1",
+      "value": "0x0"
+    }
   ],
   "transactionsRoot": "0xada45aa72d8206747ec7a2dcbaed701b921389ca9b8275a516e6fac044f23357",
   "uncles": []
