+++ optimism/testdata/goerli_bedrock_tx_5003318_1.json
@@ -1,0 +1,19 @@
+{
+    "calls": [
+        {
+            "from": "0x4200000000000000000000000000000000000015",
+            "gas": "0x8cca310",
+            "gasUsed": "0x4a28",
+            "input": "0x015d8eb900000000000000000000000000000000000000000000000000000000008097790000000000000000000000000000000000000000000000000000000063dd1a98000000000000000000000000000000000000000000000000000000000004ee2f1ed96835176d084c845bd2c09456d60401d74861b690bdabac97f6724f4b4bdf00000000000000000000000000000000000000000000000000000000000000020000000000000000000000007431310e026b69bfc676c0013e12a1a11411eec9000000000000000000000000000000000000000000000000000000000000083400000000000000000000000000000000000000000000000000000000000f4240",
+            "to": "0xc0d3c0d3c0d3c0d3c0d3c0d3c0d3c0d3c0d30015",
+            "type": "DELEGATECALL"
+        }
+    ],
+    "from": "0xdeaddeaddeaddeaddeaddeaddeaddeaddead0001",
+    "gas": "0x8f07808",
+    "gasUsed": "0xb729",
+    "input": "0x015d8eb900000000000000000000000000000000000000000000000000000000008097790000000000000000000000000000000000000000000000000000000063dd1a98000000000000000000000000000000000000000000000000000000000004ee2f1ed96835176d084c845bd2c09456d60401d74861b690bdabac97f6724f4b4bdf00000000000000000000000000000000000000000000000000000000000000020000000000000000000000007431310e026b69bfc676c0013e12a1a11411eec9000000000000000000000000000000000000000000000000000000000000083400000000000000000000000000000000000000000000000000000000000f4240",
+    "to": "0x4200000000000000000000000000000000000015",
+    "type": "CALL",
+    "value": "0x0"
+}
